moved profile from activity to new file
